chore: refactor toolkit lib.rs
