+++ .github/workflows/push.yml
@@ -19,7 +19,6 @@ jobs:
         -D clippy::cast-sign-loss
         -D clippy::checked-conversions
         -A clippy::upper-case-acronyms
-        -A special_module_name
 
     steps:
       - uses: actions/checkout@v1
