--- toolkit/src/lib/mod.rs
@@ -1,1 +1,0 @@
-pub mod data_requests;
